SemiDocAI4SPS/requirements.txt
azure-functions
azure-identity
azure-ai-documentintelligence
azure-cosmos
msal
requests
